Refactorisez les importations pour utiliser des chemins relatifs pour storage_service et mettez à jour le fichier pubspec.yaml afin d'inclure la dépendance internet_connection_checker_plus. Ajoutez l'enregistrement du plugin ConnectivityPlus dans les fichiers générés et supprimez le fichier connectivity.dart inutilisé.
