+++ pubspec.yaml
@@ -17,13 +17,12 @@ dependencies:
   shared_preferences: ^2.5.3
   flutter_staggered_animations: ^1.1.1
   percent_indicator: ^4.2.5
-
-  # local db management
   path: ^1.9.1
   sqflite_common_ffi: ^2.3.5
   sqlite3_flutter_libs: ^0.5.33
   sqflite_common_ffi_web: ^1.0.0
   path_provider: ^2.1.5
+  internet_connection_checker_plus: ^2.7.2
 
 
 dev_dependencies:
